+++ stella/vladsf/hydxsecl.f
@@ -3,8 +3,8 @@ c                 f u n c t i o n   h y d x s e c l
 c -----------------------------------------------------------------------
 c This version of HYDXSECL calls 
       function hydxsecl(phot, z, n, l)
-      implicit real(dp) (a-h,o-z)
-      real(dp) phot, z
+      implicit real*8 (a-h,o-z)
+      real*8 phot, z
       integer n, l
       logical hyphinit
